data_collection/config.py
"""
Workforce Pulse — Configuration

Montgomery-specific data collection targets aligned with hackathon insights:
  - Government (largest employer sector)
  - Defense & Air Force (Maxwell/Gunter AFB)
  - Manufacturing (Hyundai + suppliers)
  - Big Tech / Data Centers (AWS, Google, Meta expansion)
  - Public Safety (understaffed police, recruitment push)
  - Healthcare, Education (ASU, Baptist Health)
"""

import os
from pathlib import Path
from dotenv import load_dotenv

load_dotenv()

BRIGHTDATA_API_TOKEN = os.getenv("BRIGHTDATA_API_TOKEN", "")
BRIGHTDATA_UNLOCKER_ZONE = os.getenv("BRIGHTDATA_UNLOCKER_ZONE", "mcp_unlocker")
BRIGHTDATA_SERP_ZONE = os.getenv("BRIGHTDATA_SERP_ZONE", "serp_api1")

BRIGHTDATA_DC_HOST = os.getenv("BRIGHTDATA_DC_HOST", "brd.superproxy.io")
BRIGHTDATA_DC_PORT = int(os.getenv("BRIGHTDATA_DC_PORT", "33335") or "33335")
BRIGHTDATA_DC_USER = os.getenv("BRIGHTDATA_DC_USER", "")
BRIGHTDATA_DC_PASS = os.getenv("BRIGHTDATA_DC_PASS", "")

BASE_DIR = Path(__file__).resolve().parent
DATA_DIR = BASE_DIR / "data"
DATA_DIR.mkdir(exist_ok=True)

REGION = "Montgomery, AL"
GEO_LOCATION = "us"

# ── Job collection ───────────────────────────────────────────────
# Queries organized by Montgomery's actual economic sectors.

JOB_SEARCH_QUERIES = [
    # Government (dominant sector — state capital)
    "state government jobs Montgomery Alabama",
    "City of Montgomery government hiring",
    "Montgomery AL public sector employment",
    # Defense & Federal (Maxwell/Gunter AFB)
    "Maxwell Air Force Base jobs Montgomery",
    "Gunter Annex federal jobs Montgomery AL",
    "defense contractor jobs Montgomery Alabama",
    "federal technology jobs Montgomery AL",
    # Manufacturing (Hyundai anchor)
    "Hyundai manufacturing jobs Montgomery AL",
    "Montgomery AL manufacturing plant hiring",
    # Big Tech & Data Centers (emerging)
    "data center jobs Montgomery Alabama",
    "AWS Google Meta jobs Alabama",
    "technology jobs Montgomery AL",
    # Healthcare
    "healthcare jobs Montgomery Alabama",
    "Baptist Health Montgomery hiring",
    # Education
    "Alabama State University jobs Montgomery",
    "education jobs Montgomery AL",
    # Public Safety (understaffed — cross-track)
    "police officer jobs Montgomery Alabama",
    "Montgomery police department hiring",
    "public safety jobs Montgomery AL",
    # General
    "job postings Montgomery AL",
    "hiring Montgomery Alabama 2026",
]

JOB_BOARDS = [
    "https://www.indeed.com/jobs?q=&l=Montgomery%2C+AL&sort=date",
    "https://www.linkedin.com/jobs/jobs-in-montgomery-al",
    "https://jobapscloud.com/MGM/",
    "https://www.montgomerychamber.com/jobs",
    "https://www.usajobs.gov/Search/Results?l=Montgomery%2C%20Alabama",
    "https://personnel.alabama.gov/Jobs",
]

LINKEDIN_JOB_URLS = [
    "https://www.linkedin.com/jobs/search/?location=Montgomery%2C%20Alabama",
    "https://www.linkedin.com/jobs/search/?keywords=government&location=Montgomery%2C%20Alabama",
    "https://www.linkedin.com/jobs/search/?keywords=defense&location=Montgomery%2C%20Alabama",
    "https://www.linkedin.com/jobs/search/?keywords=healthcare&location=Montgomery%2C%20Alabama",
    "https://www.linkedin.com/jobs/search/?keywords=manufacturing&location=Montgomery%2C%20Alabama",
    "https://www.linkedin.com/jobs/search/?keywords=technology+data+center&location=Montgomery%2C%20Alabama",
]

# Keywords for LinkedIn SDK job search (used by linkedin_search_jobs method)
LINKEDIN_SEARCH_KEYWORDS = [
    "",  # general — all Montgomery jobs
    "government",
    "defense",
    "healthcare",
    "manufacturing",
    "technology",
    "data center",
    "police",
    "education",
]

EXTRACT_JOB_URLS = [
    (
        "https://www.indeed.com/jobs?q=&l=Montgomery%2C+AL&sort=date",
        "Return a JSON array of job listings. For each job include: "
        "title (string), company (string), location (string), salary (string or null), "
        "job_type (string: full-time/part-time/contract), posted_date (string), "
        "url (string). ONLY include actual job postings, not navigation links, "
        "headers, or UI elements. Return ONLY the JSON array, no other text.",
    ),
    (
        "https://jobapscloud.com/MGM/",
        "Return a JSON array of government job openings. For each include: "
        "title (string), department (string), salary_range (string), "
        "closing_date (string), url (string). These are City of Montgomery "
        "government positions. Return ONLY the JSON array.",
    ),
    (
        "https://www.usajobs.gov/Search/Results?l=Montgomery%2C%20Alabama",
        "Return a JSON array of federal job listings. For each include: "
        "title (string), agency (string), grade (string), salary_range (string), "
        "location (string), closing_date (string), url (string). "
        "Return ONLY the JSON array.",
    ),
]

# ── Business & economic signals ──────────────────────────────────

BUSINESS_SEARCH_QUERIES = [
    # Business filings / expansion
    "new business filings Montgomery AL 2026",
    "Montgomery Alabama business growth expansion",
    "Montgomery AL new companies opening",
    # Economic development
    "Montgomery Alabama economic development news",
    "Montgomery economic revitalization investment",
    # Data center expansion (strategic concern)
    "data center construction Montgomery Alabama",
    "AWS Google Meta data center Alabama",
    "data center impact Alabama infrastructure",
    # Defense contracting
    "defense contracts Montgomery Alabama 2026",
    "Maxwell Air Force Base contracts Montgomery",
    # Commercial real estate
    "Montgomery AL commercial real estate development",
    # Public safety
    "Montgomery police staffing shortage",
]

MONTGOMERY_COMPANIES_LINKEDIN = [
    # Government
    "https://www.linkedin.com/company/city-of-montgomery-alabama/",
    "https://www.linkedin.com/company/state-of-alabama/",
    # Defense / Federal
    "https://www.linkedin.com/company/maxwell-air-force-base/",
    # Manufacturing
    "https://www.linkedin.com/company/hyundai-motor-manufacturing-alabama/",
    # Healthcare
    "https://www.linkedin.com/company/baptist-health-montgomery/",
    "https://www.linkedin.com/company/jackson-hospital-montgomery/",
    # Education
    "https://www.linkedin.com/company/alabama-state-university/",
    "https://www.linkedin.com/company/auburn-university-at-montgomery/",
    # Big Tech (presence/expansion)
    "https://www.linkedin.com/company/amazon-web-services/",
    "https://www.linkedin.com/company/google/",
    "https://www.linkedin.com/company/meta/",
]

MONTGOMERY_COMPANIES_CRUNCHBASE = [
    "https://www.crunchbase.com/organization/hyundai-motor-manufacturing-alabama",
]

ZILLOW_COMMERCIAL_URLS = [
    "https://www.zillow.com/montgomery-al/commercial/",
]

OPEN_DATA_URLS = [
    "https://opendata.montgomeryal.gov/",
    "https://opendata.montgomeryal.gov/datasets/business-licenses",
    "https://opendata.montgomeryal.gov/datasets/building-permits",
    "https://opendata.montgomeryal.gov/datasets/code-enforcement",
]

EXTRACT_BUSINESS_URLS = [
    (
        "https://opendata.montgomeryal.gov/",
        "Extract all available datasets: name, category, description, last updated date, number of records. Focus on business licenses, permits, economic indicators, employer data, workforce, zoning, land use, public safety, and infrastructure datasets.",
    ),
    (
        "https://opendata.montgomeryal.gov/datasets/business-licenses",
        "Return a JSON array of recent business license records. For each include: "
        "business_name, license_type, address, issue_date, status. "
        "Return ONLY the JSON array.",
    ),
    (
        "https://opendata.montgomeryal.gov/datasets/building-permits",
        "Return a JSON array of recent building permit records. For each include: "
        "permit_number, project_description, address, permit_type, issue_date, "
        "estimated_cost. Return ONLY the JSON array.",
    ),
]

# ── Glassdoor employer quality ───────────────────────────────────

GLASSDOOR_EMPLOYER_URLS = [
    "https://www.glassdoor.com/Overview/Working-at-City-of-Montgomery-Alabama-EI_IE236885.htm",
    "https://www.glassdoor.com/Overview/Working-at-Hyundai-Motor-Manufacturing-Alabama-EI_IE256738.htm",
    "https://www.glassdoor.com/Overview/Working-at-Baptist-Health-EI_IE17399.htm",
    "https://www.glassdoor.com/Overview/Working-at-Alabama-State-University-EI_IE130624.htm",
    "https://www.glassdoor.com/Overview/Working-at-Maxwell-Air-Force-Base-EI_IE40073.htm",
    "https://www.glassdoor.com/Overview/Working-at-Jackson-Hospital-Montgomery-EI_IE466837.htm",
]

GLASSDOOR_SEARCH_QUERIES = [
    "Glassdoor reviews City of Montgomery Alabama employer",
    "Glassdoor reviews Hyundai Montgomery Alabama",
    "Glassdoor reviews Maxwell Air Force Base Montgomery",
    "best employers Montgomery Alabama Glassdoor ratings",
    "Montgomery AL employer reviews salary compensation",
]

# ── Google Maps local business ───────────────────────────────────

GOOGLE_MAPS_URLS = [
    "https://www.google.com/maps/place/Hyundai+Motor+Manufacturing+Alabama/@32.346,-86.296,17z",
    "https://www.google.com/maps/place/Baptist+Health+Montgomery/@32.380,-86.298,17z",
    "https://www.google.com/maps/place/Maxwell+Air+Force+Base/@32.383,-86.366,15z",
    "https://www.google.com/maps/place/Jackson+Hospital+%26+Clinic/@32.374,-86.303,17z",
    "https://www.google.com/maps/place/Alabama+State+University/@32.364,-86.296,16z",
]

GOOGLE_MAPS_SEARCH_QUERIES = [
    "top employers Montgomery Alabama",
    "new businesses Montgomery AL 2026",
    "Montgomery AL commercial district businesses",
    "Montgomery Alabama major employers reviews",
]

# ── Montgomery local education programs (for skills gap analysis) ─

LOCAL_PROGRAMS = [
    "Alabama State University",
    "Auburn University at Montgomery",
    "Troy University Montgomery",
    "Trenholm State Community College",
    "Faulkner University",
    "Huntingdon College",
]

LOCAL_DEGREE_FIELDS = [
    "nursing", "business administration", "criminal justice",
    "computer science", "information technology", "education",
    "public administration", "social work", "engineering",
    "accounting", "healthcare administration", "cybersecurity",
    "welding", "automotive technology", "hvac",
    "logistics", "supply chain", "data analytics",
]
