requirements.txt
httpx>=0.27
python-dotenv>=1.0
brightdata-sdk>=2.2
pydantic>=2.0
pytest>=7.0
pytest-asyncio>=0.21
fastapi>=0.115
uvicorn[standard]>=0.30

